# backend/docs/swagger.yaml
basePath: /api
definitions:
  backend_internal_models.AffordabilityResult:
    properties:
      avg_price:
        type: integer
      property_type:
        type: string
      relative_affordability:
        type: number
    type: object
  backend_internal_models.CSVConfigPayload:
    properties:
      bucket_key:
        type: string
      column_mapping:
        additionalProperties:
          type: string
        type: object
      file:
        description: Backward compatibility
        type: string
      has_header:
        type: boolean
      job_id:
        type: string
      path:
        description: Backward compatibility
        type: string
    type: object
  backend_internal_models.GrowthHotspotResult:
    properties:
      current_median:
        type: integer
      growth_rate:
        type: number
      prev_median:
        type: integer
      region:
        type: string
    type: object
  backend_internal_models.Job:
    properties:
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: string
      message:
        description: For error messages or success info
        type: string
      payload:
        description: Custom data the job was initiated with
        items:
          type: integer
        type: array
      progress:
        type: integer
      result:
        description: Data resulting from completion/failure
        items:
          type: integer
        type: array
      status:
        $ref: '#/definitions/backend_internal_models.JobStatus'
      task_type:
        type: string
      total:
        type: integer
      updated_at:
        type: string
    type: object
  backend_internal_models.JobStatus:
    enum:
    - PENDING
    - RUNNING
    - SUCCESS
    - FAILED
    type: string
    x-enum-varnames:
    - JobStatusPending
    - JobStatusRunning
    - JobStatusSuccess
    - JobStatusFailed
  backend_internal_models.MedianPriceResult:
    properties:
      median_price:
        type: integer
      region:
        type: string
    type: object
  backend_internal_models.NewBuildPremiumResult:
    properties:
      new_avg:
        type: integer
      old_avg:
        type: integer
      premium_percent:
        type: number
      region:
        type: string
    type: object
  backend_internal_models.PriceBracketResult:
    properties:
      bracket:
        type: string
      count:
        type: integer
      percentage:
        type: number
    type: object
  backend_internal_models.PriceTrendResult:
    properties:
      avg_price:
        type: integer
      median_price:
        type: integer
      period:
        type: string
      transaction_count:
        type: integer
    type: object
...
  backend_internal_models.PropertyTypeDistributionResult:
    properties:
      count:
        type: integer
      percentage:
        type: number
      property_type:
        type: string
    type: object
  backend_internal_models.TopActiveAreaResult:
    properties:
      region:
        type: string
      total_value:
        type: integer
      transaction_count:
        type: integer
    type: object
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  internal_routes_api.ErrorResponse:
    properties:
      error:
        type: string
      status:
        example: false
        type: boolean
    type: object
  internal_routes_api.JobListPayload:
    properties:
      data:
        items:
          $ref: '#/definitions/backend_internal_models.Job'
        type: array
      total:
        type: integer
    type: object
  internal_routes_api.JobResponse:
    properties:
      job_id:
        type: string
      message:
        type: string
      status:
        example: true
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a comprehensive API for analyzing and accessing UK housing
    market and rental data.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: UK Housing Market & Rental Intelligence API
  version: "1.0"
paths:
  /admin/jobs:
    get:
      consumes:
      - application/json
      description: Get a list of background jobs with pagination
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_routes_api.JobListPayload'
      security:
      - JwtAuth: []
      summary: List background jobs
      tags:
      - admin
  /admin/stream-upload:
    post:
      consumes:
      - application/octet-stream
      description: Stream a large CSV file directly from the request body to bucket
        and queue a migration job
      parameters:
      - description: Original filename
        in: query
        name: filename
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/internal_routes_api.JobResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_routes_api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_routes_api.ErrorResponse'
      security:
      - JwtAuth: []
      summary: Stream upload CSV file
      tags:
      - admin
  /admin/upload:
    post:
      consumes:
      - application/json
      description: Queue a background job to migrate properties from a CSV file
      parameters:
      - description: CSV Migration Configuration
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/backend_internal_models.CSVConfigPayload'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/internal_routes_api.JobResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_routes_api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_routes_api.ErrorResponse'
      security:
      - JwtAuth: []
      summary: Upload CSV for migration
      tags:
      - admin
  /admin/upload-file:
    post:
      consumes:
      - multipart/form-data
      description: Upload a CSV file via multipart/form-data to bucket and queue a
        migration job
      parameters:
      - description: CSV File
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/internal_routes_api.JobResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_routes_api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_routes_api.ErrorResponse'
      security:
      - JwtAuth: []
      summary: Upload CSV file directly
      tags:
      - admin
  /analytics/affordability:
    get:
      consumes:
      - application/json
      description: Get relative affordability by property type
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/backend_internal_models.AffordabilityResult'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_routes_api.ErrorResponse'
      summary: Get affordability index
      tags:
      - analytics
  /analytics/growth-hotspots:
    get:
      consumes:
      - application/json
      description: Get top districts with highest price growth rate
      parameters:
      - default: 10
        description: Number of results
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/backend_internal_models.GrowthHotspotResult'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_routes_api.ErrorResponse'
      summary: Get growth hotspots
      tags:
      - analytics
  /analytics/median-price:
    get:
      consumes:
      - application/json
      description: Get median price grouped by county, district, or town_city
      parameters:
      - default: county
        description: Region type (county, district, town_city)
        in: query
        name: by
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/backend_internal_models.MedianPriceResult'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_routes_api.ErrorResponse'
      summary: Get median price by region
      tags:
      - analytics
  /analytics/new-build-premium:
    get:
      consumes:
      - application/json
      description: Get average prices of new builds vs established properties by region
      parameters:
      - default: county
        description: Region type (county, district, town_city)
        in: query
        name: by
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/backend_internal_models.NewBuildPremiumResult'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_routes_api.ErrorResponse'
      summary: Get new build premium
      tags:
      - analytics
  /analytics/price-bracket-distribution:
    get:
      consumes:
      - application/json
      description: Get distribution of properties by price ranges
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/backend_internal_models.PriceBracketResult'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_routes_api.ErrorResponse'
      summary: Get price bracket distribution
      tags:
      - analytics
  /analytics/price-trend:
    get:
      consumes:
      - application/json
      description: Get average and median price trends over time
      parameters:
      - default: month
        description: Time interval (month, year)
        in: query
        name: interval
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/backend_internal_models.PriceTrendResult'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_routes_api.ErrorResponse'
      summary: Get price trend analysis
      tags:
      - analytics
  /analytics/property-type-distribution:
    get:
      consumes:
      - application/json
      description: Get distribution of properties by type (detached, semi, flat, etc.)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/backend_internal_models.PropertyTypeDistributionResult'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_routes_api.ErrorResponse'
      summary: Get property type distribution
      tags:
      - analytics
  /analytics/top-active-areas:
    get:
      consumes:
      - application/json
      description: Get regions with highest transaction volume
      parameters:
      - default: district
        description: Region type (county, district, town_city)
        in: query
        name: by
        type: string
      - default: 10
        description: Number of results
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/backend_internal_models.TopActiveAreaResult'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_routes_api.ErrorResponse'
      summary: Get top active areas
      tags:
      - analytics
...
securityDefinitions:
  JwtAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"